define mustEncode and add them to the needed places

+++ crypto/merkle/simple_proof.go
@@ -25,8 +25,8 @@ const (
 // everything.  This also affects the generalized proof system as
 // well.
 type SimpleProof struct {
-	Total    int64      `json:"total"`     // Total number of items.
-	Index    int64      `json:"index"`     // Index of item to prove.
+	Total    int64    `json:"total"`     // Total number of items.
+	Index    int64    `json:"index"`     // Index of item to prove.
 	LeafHash []byte   `json:"leaf_hash"` // Hash of item value.
 	Aunts    [][]byte `json:"aunts"`     // Hashes from leaf's sibling to a root's child.
 }
@@ -39,8 +39,8 @@ func SimpleProofsFromByteSlices(items [][]byte) (rootHash []byte, proofs []*Simp
 	proofs = make([]*SimpleProof, len(items))
 	for i, trail := range trails {
 		proofs[i] = &SimpleProof{
-			Total:    len(items),
-			Index:    i,
+			Total:    int64(len(items)),
+			Index:    int64(i),
 			LeafHash: trail.Hash,
 			Aunts:    trail.FlattenAunts(),
 		}
@@ -129,7 +129,7 @@ func (sp *SimpleProof) ToProto() *tmcrypto.SimpleProof {
 	return pb
 }
 
-func (sp *SimpleProof) FromProto(pb *tmcrypto.SimpleProof) error{
+func (sp *SimpleProof) FromProto(pb *tmcrypto.SimpleProof) error {
 	if pb == nil {
 		return errors.New("nil proof")
 	}
@@ -169,7 +169,7 @@ func (sp *SimpleProof) ValidateBasic() error {
 // Use the leafHash and innerHashes to get the root merkle hash.
 // If the length of the innerHashes slice isn't exactly correct, the result is nil.
 // Recursive impl.
-func computeHashFromAunts(index int, total int, leafHash []byte, innerHashes [][]byte) []byte {
+func computeHashFromAunts(index int64, total int64, leafHash []byte, innerHashes [][]byte) []byte {
 	if index >= total || index < 0 || total <= 0 {
 		return nil
 	}
@@ -243,7 +243,7 @@ func trailsFromByteSlices(items [][]byte) (trails []*SimpleProofNode, root *Simp
 		trail := &SimpleProofNode{leafHash(items[0]), nil, nil, nil}
 		return []*SimpleProofNode{trail}, trail
 	default:
-		k := getSplitPoint(len(items))
+		k := getSplitPoint(int64(len(items)))
 		lefts, leftRoot := trailsFromByteSlices(items[:k])
 		rights, rightRoot := trailsFromByteSlices(items[k:])
 		rootHash := innerHash(leftRoot.Hash, rightRoot.Hash)
